backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pymongo==4.6.0
psycopg2==2.9.9
mysql-connector-python==8.2.0
supabase==2.3.0